+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Site setting to restrict changing system and branch after signup
 - Ability to create groups of challenges with unique URLs
 - Ability to relate challenges and challenge groups to events
+- Site setting to require events to be created with badges
 
 ### Fixed
 - Issue with favorites list filtering